feat: add support for saving snapshot history to files
We're going to add support for saving snapshot history to files. There are a few parts to this. First, we need to write out the files to a directory with a name. The format of the name will be year, month, date, dash, slug, where slug is computed by Claude's understanding of the title. We don't have code for that yet, by the way, so we'll have to add it. And we need a small SQLite database which maps the UUID from the chat URL to this. So the overall flow will be when we want to save a snapshot history. Instead of saving it to a file, we will first look up the UUID, see if there's already a file for it. And then if there is not one, we will make a new file name, put it in the SQLite database, and write a file. The recordings will be saved to CWD under a directory name of your choosing.

```git-revs
a470b99  (Base revision)
120e789  Add SQLite as a dependency
fc8b13a  Create storage module for snapshot management
14fa7ff  Add imports for the storage module
09c3771  Add function to extract chat UUID from URL
d5a980e  Update run_snapshot_history to use storage system
752073f  Update CLI options and main function for snapshot storage
226c340  Fix legacy snapshot file handling
e1b4d3b  Add snapshot history to features in README
2260de3  Add snapshot usage instructions to README
7e8ed5f  Snapshot before codemcp change
785374e  Remove legacy snapshot file flag from CLI options
777492d  Remove deprecated snapshot file warning code
e8316d8  Remove legacy snapshot file handling code
82f7154  Remove try-except from _init_db to propagate errors
c1430e3  Remove try-except from update_snapshot to propagate errors
b691981  Remove try-except from list_snapshots to propagate errors
ae74d02  Remove try-except from storage initialization in CLI
HEAD     Remove try-except from run_snapshot_history function
```

codemcp-id: 10-feat-add-support-for-saving-snapshot-history-to-fi