test: Add test for not auto-continuing with unrelated textarea content
testdata/20250410/hit_max_length_textarea_other.xml is a test file for the situation where we've hit max length and an auto continue potentially could occur, but in the XML, because the textarea has unrelated text in it, we actually should NOT trigger an autocontinue. Use this XML to create a test case similar to the existing ones and assert auto continue does NOT trigger. Do NOT write a new XML, the XML is already provided. Do NOT read the xml code, you don't need to know it to do this task.

```git-revs
5a213d5  (Base revision)
HEAD     Create test file for checking auto-continue with unrelated textarea content
```

codemcp-id: 197-test-add-test-for-not-auto-continuing-with-unrelat