feat: add test for notify_on_complete with XML swapping
in_progress.xml and in_progress2.xml are testdata/20250410/ a pair of XML files. In the first one, chat is in progress. In the second, the in progress generation has completed. We want to use these test files to drive a test to see that notify_on_complete feature works. However, there is a complication: we need to run a function with the first XML, and then when we're done run it again with the second XML. There isn't any precedent for this in the existing code, so you may need to refactor so it's possible to swap out the XML, simulating the HAX state getting updated.

```git-revs
0abba2e  (Base revision)
d090924  Create test file for notify_on_complete feature with XML swapping
c3bb695  Add debug logging and import check_chat_running_state function
71455b9  Add detailed debug methods to test file
948d680  Create test file with mocking for notify_on_complete feature
a709ecc  Add documentation for mock-based testing approach
HEAD     Update test_notify_on_complete.py to use mocking instead of XML swapping
```

codemcp-id: 198-feat-add-test-for-notify-on-complete-with-xml-swap