feat: add Chrome extension for auto-clicking modal allow button
In packages/extension create a TS package in our monorepo for a Chrome extension. This Chrome extension does a PageStateMatcher for .z-modal and if it appears it clicks the "Allow for this chat" button in the modal dialog.

To test, we have already setup a test case packages/rrweb-headless/src/index.test.ts which has a rrweb recording that should trigger this matcher in the test "should launch a real browser and replay events". We should modify this test to load our Chrome extension in the page context. You will need some extensibility points to make sure the extension is testable; specifically, we need a way to stub out the "click" action and instead just report something to the DOM so we can match it from playwright. One way to do this is to have the extension test if it's operating on the rrweb replay player, and toggle between the behaviors if it sees it's on it (it needs to test this anyway so that it can write results appropriately).

```git-revs
cdb86c2  (Base revision)
8ba3ae0  Creating package.json for Chrome extension
74fa8c0  Creating tsconfig.json for Chrome extension
ca32195  Creating manifest.json for Chrome extension
abe4326  Creating background script for Chrome extension
a5b0749  Creating content script for Chrome extension
849c203  Creating rrweb player utilities for testing
b982cc2  Creating index file for utils
52bba1e  Creating vitest config for extension
98bb37c  Creating test file for extension
1a21a3c  Updating test to integrate Chrome extension
515d642  Adding chromiumArgs option to RrwebReplayOptions
5024206  Adding page property to RrwebReplayResult
33ad5c7  Updating runRrwebReplay to use chromiumArgs and include page in result
15708bc  Adding extension as a dependency to rrweb-headless
de0ea79  Creating README for Chrome extension
HEAD     Updating test with browser cleanup
```

codemcp-id: 25-feat-add-chrome-extension-for-auto-clicking-modal-