feat: add Chrome extension for auto-clicking modal allow button
In packages/extension create a TS package in our monorepo for a Chrome extension. This Chrome extension does a PageStateMatcher for .z-modal and if it appears it clicks the "Allow for this chat" button in the modal dialog.

To test, we have already setup a test case packages/rrweb-headless/src/index.test.ts which has a rrweb recording that should trigger this matcher in the test "should launch a real browser and replay events". We should modify this test to load our Chrome extension in the page context. You will need some extensibility points to make sure the extension is testable; specifically, we need a way to stub out the "click" action and instead just report something to the DOM so we can match it from playwright. One way to do this is to have the extension test if it's operating on the rrweb replay player, and toggle between the behaviors if it sees it's on it (it needs to test this anyway so that it can write results appropriately).

```git-revs
7592af8  (Base revision)
c61e038  Update build script to copy manifest.json to dist directory
1fa8517  Snapshot before auto-test
6f405d5  Fix TypeScript config for extension package
294b94f  Update build script to skip dts generation and ensure dist directory exists
HEAD     Remove the problematic dts flag from build script
```

codemcp-id: 25-feat-add-chrome-extension-for-auto-clicking-modal-