fix(extension): handle dynamic tracking button in responseStateObserver
The logic in packages/extension/src/observers/responseStateObserver.ts is problematic. Specifically, empirically (as well as in the test) we observe the button "respawning" (multiple occurrences of Tracking button in state). Rewrite it so that it doesn't assume the DOM element for the button always sticks around but if we requery for the button we will be able to find it and it has the correct state.

```git-revs
a26806a  (Base revision)
762a3c2  Snapshot before codemcp change
811738b  Rewrite responseStateObserver to query DOM instead of tracking specific button element
HEAD     Update test expectation for response state log message
```

codemcp-id: 77-fix-extension-handle-dynamic-tracking-button-in-re