test: add test for avoiding auto-continue on navigate
testdata/navigate-to-continue-page.json contains a test file which we will use to verify that, when auto-continue is enabled for our extension, we do NOT immediately trigger the continue in this test. So we should have a test somewhat similar to test/index.test.ts 'should detect the Continue button in hit-max-length-may-2025.json' but we should NOT have detected a continue button coming into existence.

```git-revs
2fcf159  (Base revision)
HEAD     Add test case for navigate-to-continue-page.json that verifies we don't auto-continue when navigating to a page with a Continue button
```

codemcp-id: 92-test-add-test-for-avoiding-auto-continue-on-naviga